pyproject.toml
[tool.poetry]
name = "insights"
version = "0.1.0"
description = ""
authors = ["Amer Alnajdawi <amer.alnajdawi@treetop-medical.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.10.14"
streamlit = "^1.35.0"
streamlit-chat = "^0.1.1"
pdfplumber="0.9.0"
pandas="2.2.2"
openai="1.41.0"
pydantic = "^2.6.1"
python-dotenv = "^1.0.1"




[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
